+++ 2D_SPH_from_begining/classes/Vector2D.hpp
@@ -12,11 +12,11 @@
 class Vector2D {
 public:
     Vector2D() : x(0), y(0) {}
-    Vector2D(float x) : x(x), y(x) {}
-    Vector2D(float x, float y) : x(x), y(y) {}
+    Vector2D(double x) : x(x), y(x) {}
+    Vector2D(double x, double y) : x(x), y(y) {}
     ~Vector2D() {}
 
-    float       x, y;
+    double       x, y;
 
     Vector2D    operator+(const Vector2D& a) const {
         return (Vector2D(x + a.x, y + a.y));
@@ -26,19 +26,19 @@ public:
         return (Vector2D(x - a.x, y - a.y));
     }
 
-    Vector2D    operator*(const float a) const {
+    Vector2D    operator*(const double a) const {
         return (Vector2D(x * a, y * a));
     }
 
-    Vector2D    operator/(const float a) const {
+    Vector2D    operator/(const double a) const {
         return (Vector2D(x / a, y / a));
     }
 
-    Vector2D    operator+=(const float a) const {
+    Vector2D    operator+=(const double a) const {
         return (Vector2D(x + a, y + a));
     }
 
-    Vector2D    operator-=(const float a) const {
+    Vector2D    operator-=(const double a) const {
         return (Vector2D(x - a, y - a));
     }
 
@@ -52,7 +52,7 @@ public:
         this->y -= a.y;
     }
 
-    float       dot(const Vector2D& a) const {
+    double       dot(const Vector2D& a) const {
         return (x * a.x + y * a.y);
     }
 };
